requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-jwt
python-decouple
mysqlclient
requests
python-dotenv
google-generativeai
Pillow